+++ test/sandbox/app/components/invalid_with_render_component.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-class InvalidWithRenderComponent < ViewComponent::Base
-  # You aren't permitted to implement this method.
-  def with_content
-  end
-end